requirements.txt

aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.43
botocore==1.35.43
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
coloredlogs==15.0.1
contourpy==1.3.0
crewai==0.74.0
crewai-tools==0.13.1
cryptography==43.0.1
cycler==0.12.1
databricks-sdk==0.35.0
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.123
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.2
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.5.2
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json_repair==0.30.0
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.14.0
langchain==0.3.3
langchain-cohere==0.3.1
langchain-community==0.3.2
langchain-core==0.3.12
langchain-experimental==0.3.2
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.136
litellm==1.49.7
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.19
mlflow==2.17.0
mlflow-skinny==2.17.0
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
neo4j==5.25.0
networkx==3.4.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.52.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pathos==0.3.3
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.18.2
pyparsing==3.2.0
pypdf==5.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.385
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.12.0
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
setuptools==75.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
tomli_w==1.1.0
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.4.24
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.15.4
zipp==3.20.2
